[PHASE-10.2] IMPROVE: Add cast safety documentation and fix getleader() type signature
Analyzed HIGH RISK integer/long casts flagged by analyze_int_long_casts.py.
All analyzed casts were FALSE POSITIVES - safe widening operations.

Cast Safety Documentation (7 comments added):
- trade.c: 3 comments for safe int→long widening (lines 182, 1228, 1754)
- npc.c: 1 comment for safe min() return widening (line 1048)
- commands.c: 1 comment for safe coordinate widening (line 1507)
- move.c: 2 comments for safe coordinate arithmetic widening (lines 275, 284)

Type Safety Fix (getleader function):
- misc.c:3430: Changed parameter from int to short (matches s_nation.class type)
- data.h:1030: Updated prototype to accept short instead of int
- update.c:952: Removed unnecessary (int) cast from getleader() call

Pattern Identified:
- (long) casts: Function return values widened from int→long (always safe)
- (int) casts: Coordinates widened from unsigned char→int (0-255 range, always safe)

Analysis Tool Limitation:
Regex pattern detects any cast on lines with array syntax, not specifically
casts used FOR array indexing. This created high false positive rate (100%).

Verified: All modified files compile with zero warnings at level 9.

+++ data.h
@@ -1027,7 +1027,7 @@ extern void get_nname(char str[]);
 extern int get_country(void);
 extern int get_god(void);
 extern void reset_god(void);
-extern int getleader(int class);
+extern int getleader(short class);
 extern int mailopen(int to);
 extern void mailclose(int to);
 extern int markok(int mark, int prtflag);
